Move benchmarks to separate file
Also, remove BSL header.
