+++ benchmark_test.go
@@ -1,0 +1,45 @@
+package main
+
+import "testing"
+
+const (
+	numRows = 4096
+	numCols = 1
+)
+
+func BenchmarkRowBasedInterface(b *testing.B) {
+	scan := &tableReader{rows: makeInput(numRows, numCols, Int{})}
+	render := mulOperator{
+		input:             scan,
+		arg:               Int{2},
+		columnsToMultiply: []int{0},
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		for {
+			row := render.next()
+			if row == nil {
+				break
+			}
+		}
+		scan.reset()
+	}
+}
+
+func mulInt(a, b Int) Int {
+	return Int{int: a.int * b.int}
+}
+
+func BenchmarkSpeedOfLight(b *testing.B) {
+	rows := make([]Int, numRows)
+	for i := range rows {
+		rows[i].int = i
+	}
+	arg := Int{int: 2}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		for j := range rows {
+			_ = mulInt(rows[j], arg)
+		}
+	}
+}
