+++ row_based_interface.go
@@ -1,16 +1,4 @@
-// Copyright 2019 The Cockroach Authors.
-//
-// Use of this software is governed by the Business Source License
-// included in the file licenses/BSL.txt.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0, included in the file
-// licenses/APL.txt.
-
 package main
-
-import "testing"
 
 func next() []Datum {
 	return nil
@@ -24,6 +12,7 @@ type Datum interface {
 	Mul(Datum) Datum
 }
 
+// Int implements the Datum interface.
 type Int struct {
 	int
 }
@@ -33,6 +22,7 @@ func (i Int) Mul(datum Datum) Datum {
 	return Int{int: i.int * arg}
 }
 
+// Float64 implements the Datum interface.
 type Float64 struct {
 	float64
 }
@@ -101,46 +91,4 @@ func makeInput(numRows int, numCols int, t Datum) [][]Datum {
 		panic("unhandled type")
 	}
 	return result
-}
-
-const (
-	numRows = 4096
-	numCols = 1
-)
-
-func BenchmarkInterface(b *testing.B) {
-	scan := &tableReader{rows: makeInput(numRows, numCols, Int{})}
-	render := mulOperator{
-		input:             scan,
-		arg:               Int{2},
-		columnsToMultiply: []int{0},
-	}
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		for {
-			row := render.next()
-			if row == nil {
-				break
-			}
-		}
-		scan.reset()
-	}
-}
-
-func mulInt(a, b Int) Int {
-	return Int{int: a.int * b.int}
-}
-
-func BenchmarkSpeedOfLight(b *testing.B) {
-	rows := make([]Int, numRows)
-	for i := range rows {
-		rows[i].int = i
-	}
-	arg := Int{int: 2}
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		for j := range rows {
-			_ = mulInt(rows[j], arg)
-		}
-	}
 }
